feat: JWST-Pillars Theme, 320x480 Backgrounds for existing themes (#54)
* Added 320x480 Background Images

Added 320x480 sized PNG backgrounds of every theme for the NV14 display. Used FastStone Photo Resizer for the bulk of images, and Photoshop for others that needed modification. Also added all Photoshop files to a separate folder.

* Added JWST-Pillars Theme

Created a theme with the JWST Pillars of Creation image.

* JWST-Pillars added to README

Added the JWST-Pillars theme to the list of available themes in the README.md file

* Renumbered screenshots and updated logo

* JWST-Pillars added Screenshot, Updated README

* README Hyperlink Addition

Added a hyperlink to the EdgeTX Companion text to make finding it easier.